feat: ENTTRIBE-39817: introduce rockylinux9 in matrix builds
BREAKING CHANGE: Though this is strictly a utility that is not breaking, many if not most of the use cases of using this utility will break if no adapations on it's use are made. This feature should as such be accompanied (if required) with changes in your repository to support rockylinux 9.